# app/models.py
# generated by fastapi-codegen:
#   filename:  openapi/smg_api.yaml
#   timestamp: 2024-07-22T07:59:18+00:00

from __future__ import annotations

from typing import Optional

from pydantic import BaseModel, Field


class Sentence(BaseModel):
    id: Optional[int] = Field(None, example=10)
    text: Optional[str] = Field(
        None, description='text contained in the sentence', example='super movie title'
    )


class SentenceWithCypher(BaseModel):
    id: Optional[int] = Field(None, example=10)
    text: Optional[str] = Field(
        None, description='text contained in the sentence', example='super movie title'
    )
    cyphered_text: Optional[str] = Field(
        None, description='cyphered text with rot13', example='fhcre zbivr gvgyr'
    )